Fix LDAP login CI assertions
